Se agrega excepciones, dto y controlador

+++ src/main/java/com/ibm/academia/restapi/repositorios/ApuestaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.ibm.academia.restapi.modelos.entidades.Apuesta;
 
 @Repository
-public interface ApuestaRepository extends CrudRepository<Apuesta, Integer>
+public interface ApuestaRepository extends CrudRepository<Apuesta, Long>
 {
 
 }
