+++ src/main/java/com/ibm/academia/restapi/repositorios/RuletaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ibm.academia.restapi.modelos.entidades.Ruleta;
 
-public interface RuletaRepository extends CrudRepository<Ruleta, Integer>
+public interface RuletaRepository extends CrudRepository<Ruleta, Long>
 {
 
 }
