+++ src/main/java/com/ibm/academia/restapi/servicios/GenericoDAO.java
@@ -4,8 +4,8 @@ import java.util.Optional;
 
 public interface GenericoDAO<Entidad>
 {
-    public Optional<Entidad> buscarPorId(Integer id);
+    public Optional<Entidad> buscarPorId(Long id);
     public Entidad guardar(Entidad entidad);
     public Iterable<Entidad> mostrarTodos();
-    public void eliminar(Integer id);
+    public void eliminar(Long id);
 }
