+++ src/main/java/com/ibm/academia/restapi/servicios/RuletaDAO.java
@@ -8,6 +8,6 @@ import com.ibm.academia.restapi.modelos.entidades.Ruleta;
 public interface RuletaDAO extends GenericoDAO<Ruleta>
 {
     public Boolean abrirRuleta(Ruleta ruleta);
-    public Integer guardarRuleta(Ruleta ruleta);
+    public Long guardarRuleta(Ruleta ruleta);
     public List<Apuesta> cerrarRuleta(Ruleta ruleta);
 }
